Maximum Element
This question is a bit tricky as applying simple stack solution would not satisfy all the test cases. If you will apply the simple stack technique, then there will be two cases that will show TLE and your solution will not be accepted. So, I figured out this solution as per the sole demand of the question and got all the test cases passed. Hope you all get it.
Thank You.