Largest Rectangle
This is one of the interesting questions which is based on stack. This can be solved by greedy algorithm but that would be very complex . That's where Stack comes to rescue here. 
A video link to understand what this question is all about : https://www.youtube.com/watch?v=VNbkzsnllsU