Solve Me First
I know this question seems easy, but I have provided the solution to almost all questions be it easy or hard so that beginners can understand how to bring out the output according to the desired output pattern (the pattern in which the output is desired). Also, I saw many beginners don't understand the concept of test cases and don't put an extra loop which will handle number of test cases. So, my solutions may help them to get an idea about it.  THANK YOU